pyproject.toml

[tool.nonebot]
plugins = [
    "nonebot_plugin_apscheduler",
    "nonebot_plugin_htmlrender",
    "nonebot_plugin_alconna",
    "nonebot_plugin_localstore",
    "nonebot_plugin_orm",
]
plugin_dirs = ["src/plugins"]
builtin_plugins = []
[[tool.nonebot.adapters]]
name = "OneBot V11"
module_name = "nonebot.adapters.onebot.v11"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["F", "W", "E", "UP", "ASYNC", "C4", "T10", "PYI", "PT", "Q", "RUF"]
ignore = ["E402", "E501", "UP037", "RUF001", "RUF002", "RUF003"]

[tool.pyright]
typeCheckingMode = "basic"

[project]
requires-python = "<4.0,>=3.10"
dependencies = [
    "nonebot2[fastapi,aiohttp,websockets]>=2.4.1",
    "nonebot-plugin-orm[mysql]>=0.8.1",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot-plugin-htmlrender==0.4.0",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-userinfo>=0.2.6",
    "aiohttp>=3.10.0",
    "ujson>=5.10.0",
    "nonebot-plugin-alconna>=0.55.0",
    "aiocache>=0.12.3",
    "cookit>=0.8.1",
    "pyncm>=1.6.17",
    "nonebot-plugin-localstore>=0.7.1",
    "async-lru>=2.0.4",
    "psutil>=6.1.0",
    "py-cpuinfo>=9.0.0",
    "setuptools>=75.8.0",
    "pil-utils>=0.1.12",
    "httpx==0.27.2",
    "tortoise-orm[accel,asyncmy]>=0.24.2",
    "openai>=1.76.0",
    "tomlkit>=0.13.2",
    "nb-cli>=1.4.2",
]
name = "u1bot"
version = "0.1.0"
description = ""
readme = "README.md"

[dependency-groups]
dev = ["pyright>=1.1.402"]
